+++ src/Arlen/MVC/Middleware/ALNSecurityHeadersMiddleware.m
@@ -7,29 +7,11 @@
 @interface ALNSecurityHeadersMiddleware ()
 
 @property(nonatomic, copy) NSString *contentSecurityPolicy;
+@property(nonatomic, copy) NSDictionary<NSString *, NSString *> *resolvedResponseHeaders;
 
 @end
 
-@implementation ALNSecurityHeadersMiddleware
-
-- (instancetype)initWithContentSecurityPolicy:(NSString *)contentSecurityPolicy {
-  self = [super init];
-  if (self) {
-    _contentSecurityPolicy =
-        [contentSecurityPolicy copy] ?: @"default-src 'self'";
-  }
-  return self;
-}
-
-- (void)ensureHeader:(NSString *)name value:(NSString *)value response:(ALNResponse *)response {
-  if ([response headerForName:name] == nil) {
-    [response setHeader:name value:value];
-  }
-}
-
-- (BOOL)processContext:(ALNContext *)context error:(NSError **)error {
-  (void)error;
-  ALNResponse *response = context.response;
+static NSDictionary<NSString *, NSString *> *ALNSecurityHeadersDefaults(void) {
   static NSDictionary<NSString *, NSString *> *defaults = nil;
   static dispatch_once_t defaultsOnce;
   dispatch_once(&defaultsOnce, ^{
@@ -42,10 +24,32 @@
       @"X-Permitted-Cross-Domain-Policies" : @"none",
     };
   });
-  [response setHeadersIfMissing:defaults];
-  if ([self.contentSecurityPolicy length] > 0) {
-    [self ensureHeader:@"Content-Security-Policy" value:self.contentSecurityPolicy response:response];
+  return defaults;
+}
+
+@implementation ALNSecurityHeadersMiddleware
+
+- (instancetype)initWithContentSecurityPolicy:(NSString *)contentSecurityPolicy {
+  self = [super init];
+  if (self) {
+    _contentSecurityPolicy =
+        [contentSecurityPolicy copy] ?: @"default-src 'self'";
+    NSMutableDictionary *headers = [ALNSecurityHeadersDefaults() mutableCopy];
+    if ([_contentSecurityPolicy length] > 0) {
+      headers[@"Content-Security-Policy"] = _contentSecurityPolicy;
+    }
+    _resolvedResponseHeaders = [headers copy];
   }
+  return self;
+}
+
+- (NSDictionary<NSString *, NSString *> *)responseHeaders {
+  return self.resolvedResponseHeaders;
+}
+
+- (BOOL)processContext:(ALNContext *)context error:(NSError **)error {
+  (void)error;
+  [context.response setHeadersIfMissing:[self responseHeaders]];
   return YES;
 }